Cleaning up alembic migrations mistakes
