--- api/alembic/versions/4861aa281864_added_blacklisted_token_model.py
@@ -1,30 +1,0 @@
-"""Added blacklisted token model
-
-Revision ID: 4861aa281864
-Revises: 415d1e1368bb
-Create Date: 2024-03-07 17:16:21.223769
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '4861aa281864'
-down_revision: Union[str, None] = '415d1e1368bb'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
