--- api/alembic/versions/7eba7b99a7c1_added_blacklisted_token_model_with_.py
@@ -1,30 +1,0 @@
-"""Added blacklisted token model, with columns this time
-
-Revision ID: 7eba7b99a7c1
-Revises: 4861aa281864
-Create Date: 2024-03-07 17:29:29.963757
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '7eba7b99a7c1'
-down_revision: Union[str, None] = '4861aa281864'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
